Claude_projects/owner_statement_mvp/requirements.txt
python-dotenv>=1.0.1
pyyaml>=6.0.2
requests>=2.32.3
openpyxl>=3.1.5
pandas>=2.2.2
